pyproject.toml
[tool.poetry]
name = "pelican-optimise-images"
version = "0.1.0"
description = ""
authors = ["Kevin Houlihan <kevin@crimsoncookie.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.2.0"
click = "^8.0.1"
beautifulsoup4 = "4.8.2"
lxml = "4.2.1"
pelican = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
optimise_image = "main:optimise"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
